# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервис для автоматизации рекрутинга.",
        "title": "AI Recruiting API",
        "contact": {},
        "version": "1.0"
    },
    "host": "ai-recruiting.onrender.com",
    "basePath": "/",
    "paths": {
        "/applications": {
            "post": {
                "description": "Создает отклик и запись в RESUME_DATA (ИИ-анализ)",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portal"
                ],
                "summary": "Подача отклика",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID кандидата",
                        "name": "candidate_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID вакансии",
                        "name": "vacancy_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешный отклик и ID заявки",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/applications/{id}/ai-data": {
            "get": {
                "description": "Получает вердикт, скоринг и распознанный текст из RESUME_DATA",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recruiter_tools"
                ],
                "summary": "Результаты анализа ИИ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID заявки (application_id)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные анализа",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/applications/{id}/status": {
            "patch": {
                "description": "Обновляет статус конкретного отклика (например: 'Интервью', 'Оффер', 'Отказ') по его ID в URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recruiter_tools"
                ],
                "summary": "Изменение статуса заявки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID заявки (application_id)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус для установки",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус успешно обновлен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка: Неверный формат JSON",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка: Проблема при обновлении в базе данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/my-applications": {
            "get": {
                "description": "Получение истории всех откликов текущего кандидата",
                "tags": [
                    "portal"
                ],
                "summary": "Мои отклики",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID кандидата",
                        "name": "candidate_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
...
        "/templates/{id}/generate": {
            "post": {
                "description": "Получает шаблон сообщения по ID, заменяет в нем плейсхолдеры {name} и {job} данными кандидата и формирует прямую ссылку на чат.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recruiter_tools"
                ],
                "summary": "Генерация сообщения для Telegram",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID шаблона сообщения в базе данных",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для генерации сообщения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "candidate_name": {
                                    "type": "string"
                                },
                                "telegram_username": {
                                    "type": "string"
                                },
                                "vacancy_title": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пример ответа: { 'generated_text': '...', 'telegram_link': '...' }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка: Неверные входные данные или пустые поля",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка: Шаблон с таким ID не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/vacancies/link/{short_link}": {
            "get": {
                "description": "Получение деталей вакансии через short_link",
                "tags": [
                    "portal"
                ],
                "summary": "Просмотр вакансии по ссылке",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Короткая ссылка вакансии",
                        "name": "short_link",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Vacancy"
                        }
                    },
                    "404": {
                        "description": "Вакансия не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Vacancy": {
            "type": "object",
            "properties": {
                "ai_filters": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_archived": {
                    "type": "boolean"
                },
                "recruiter_id": {
                    "type": "string"
                },
                "short_link": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}